에러 해결 중
ProcessMove(NPC)시 호출 당시엔 NPC의 주솟값은 0x00001234처럼 정상, 호출 후 NPC 주소값이 0x00101234...로 변경되어있음

1. accept시 new 하여 할당하는것이아닌 미리 할당하고 재사용
2. disconnet 의심
3.

+++ SERVER/GSP_Termproject/Manager/SectorMgr.cpp
@@ -5,7 +5,7 @@
 
 void SectorMgr::Insert(Client* Target)
 {
-	Sector& SectorSet = Sectors[Target->Position.Y / SECTORSIZE][Target->Position.X / SECTORSIZE];
+	Sector& SectorSet = Sectors[(int)Target->Position.Y / SECTORSIZE][(int)Target->Position.X / SECTORSIZE];
 
 	SectorSet.SectorLock.lock();
 	SectorSet.SectorClient.insert(Target);
@@ -14,13 +14,13 @@ void SectorMgr::Insert(Client* Target)
 
 void SectorMgr::UnsafeInsert(Client* Target)
 {
-	Sector& SectorSet = Sectors[Target->Position.Y / SECTORSIZE][Target->Position.X / SECTORSIZE];
+	Sector& SectorSet = Sectors[(int)Target->Position.Y / SECTORSIZE][(int)Target->Position.X / SECTORSIZE];
 	SectorSet.SectorClient.insert(Target);
 }
 
 void SectorMgr::Remove(Client* Target)
 {
-	Sector& SectorSet = Sectors[Target->Position.Y / SECTORSIZE][Target->Position.X / SECTORSIZE];
+	Sector& SectorSet = Sectors[(int)Target->Position.Y / SECTORSIZE][(int)Target->Position.X / SECTORSIZE];
 
 	SectorSet.SectorLock.lock();
 	SectorSet.SectorClient.erase(Target);
